В Article пусту картинку відображає, з бд завантажує та показує на сайті. До 1.100

+++ yiiProject/models/ImageUpload.php
@@ -11,33 +11,36 @@ class ImageUpload extends Model
 
     public $image;
 
-//    public function rules(){
-//        return[
-//            [['image'], 'required'],
-//            [['image'],'file', 'extensions' => 'jpg,png']
-//        ];
-//    }
-//
-//
+    public function rules(){
+        return[
+            [['image'], 'required'],
+            [['image'],'file', 'extensions' => 'jpg,png']
+        ];
+    }
+
+
     public  function uploadFile(UploadedFile $file, $currentImage)
     {
         $this->image = $file;
 
-        unlink(Yii::getAlias('@web') . 'uploads/' . $currentImage);
+        if ($this->validate()) {
 
-        $filename = strtolower(md5(uniqid($file->baseName)) . '.' . $file->extension);
+            $this->deleteCurrentImage($currentImage);
 
-        $file->saveAs(Yii::getAlias('@web') . 'uploads/' . $filename);
+            $filename = strtolower(md5(uniqid($file->baseName)) . '.' . $file->extension);
 
-        return $filename;
+            $file->saveAs(Yii::getAlias('@web') . 'uploads/' . $filename);
 
+
+            return $filename;
+        }
     }
 
-//    public function deleteCurrentImage($currentImage)
-//    {
-//        if (file_exists(Yii::getAlias('@web') . 'uploads/'  . $currentImage) &&
-//            is_file(Yii::getAlias('@web') . 'uploads/'  . $currentImage)) {
-//            unlink(Yii::getAlias('@web') . 'uploads/'  . $currentImage);
-//        }
-//    }
+    public function deleteCurrentImage($currentImage)
+    {
+        if (file_exists(Yii::getAlias('@web') . 'uploads/'  . $currentImage) &&
+            is_file(Yii::getAlias('@web') . 'uploads/'  . $currentImage)) {
+            unlink(Yii::getAlias('@web') . 'uploads/'  . $currentImage);
+        }
+    }
 }
